+++ src/core/alerts/clear-search-history-alert.ts
@@ -1,9 +1,11 @@
+import { Localization } from '@core/localize/localize.models'
 import { Alert } from 'react-native'
 
 export const clearSearchHistoryAlert = (
+  localize: () => Localization,
   onPress: () => void,
 ) => {
-  Alert.alert('Are you sure?', 'Do you really want to clear your search history?', [
+  Alert.alert(localize().are_you_sure, localize().clear_search_history_confirmation, [
     {
       text: 'Cancel',
       onPress: () => null,