+++ src/core/localize/localize.models.ts
@@ -1,9 +1,12 @@
 export interface Localization {
   premium: 'premium',
+  are_you_sure: 'are_you_sure',
+  clear_search_history_confirmation: 'clear_search_history_confirmation',
   enter_premium_code: 'enter_premium_code',
   haptic_feedback: 'haptic_feedback',
   native_search_engine: 'native_search_engine',
   clear_search_history: 'clear_search_history',
+  search_history: 'search_history',
   advanced_settings: 'advanced_settings',
   settings: 'settings',
   app_version: 'app_version',